Extract component for theme.
